requirements.txt
kubernetes==9.0.0
apache-airflow[crypto]
boto3
